# apps/core-api/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/purchase-invoices": {
      "post": {
        "operationId": "PurchaseInvoiceController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePurchaseInvoiceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "PurchaseInvoice"
        ]
      },
      "get": {
        "operationId": "PurchaseInvoiceController_findAll",
        "parameters": [
          {
            "name": "vendorId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "PurchaseInvoice"
        ]
      }
    },
    "/api/purchase-invoices/{id}": {
      "get": {
        "operationId": "PurchaseInvoiceController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "PurchaseInvoice"
        ]
      },
      "patch": {
        "operationId": "PurchaseInvoiceController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePurchaseInvoiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "PurchaseInvoice"
        ]
      },
      "delete": {
        "operationId": "PurchaseInvoiceController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PurchaseInvoice"
        ]
      }
    },
    "/api/purchase-invoices/{id}/post": {
      "patch": {
        "operationId": "PurchaseInvoiceController_post",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "PurchaseInvoice"
        ]
      }
    },
    "/api/purchase-invoices/{id}/pay": {
      "patch": {
        "operationId": "PurchaseInvoiceController_pay",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "PurchaseInvoice"
        ]
      }
    },
    "/api/purchase-invoices/{id}/lines/{lineId}": {
      "delete": {
        "operationId": "PurchaseInvoiceController_removeLine",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lineId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PurchaseInvoice"
        ]
      }
    },
...
    "/api/sales/invoices/{id}": {
      "get": {
        "operationId": "SalesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Sales"
        ]
      }
    },
...
    "/api/labor/search": {
      "get": {
        "operationId": "LaborController_search",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workshopOrderId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LaborOperationSearchResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "labor"
        ]
      }
    },
    "/api/labor/operations": {
      "get": {
        "operationId": "LaborController_listOperations",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term matching code or description",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "description": "Filter by category ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Filter by active status",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "type": "number"
            }
          },
          {
            "name": "sortField",
            "required": false,
            "in": "query",
            "description": "Sort field",
            "schema": {
              "type": "string",
              "enum": [
                "code",
                "description",
                "standardAw",
                "hourlyRate",
                "createdAt"
              ]
            }
          },
          {
            "name": "sortDirection",
            "required": false,
            "in": "query",
            "description": "Sort direction",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedLaborOperationsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "labor"
        ]
      },
      "post": {
        "operationId": "LaborController_createOperation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLaborOperationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LaborOperationResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "labor"
        ]
      }
    },
    "/api/labor/operations/{id}": {
      "get": {
        "operationId": "LaborController_getOperation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LaborOperationResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "labor"
        ]
      },
      "patch": {
        "operationId": "LaborController_updateOperation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLaborOperationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LaborOperationResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "labor"
        ]
      },
      "delete": {
        "operationId": "LaborController_removeOperation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SoftDeleteResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "labor"
        ]
      }
    },
    "/api/labor/categories": {
      "get": {
        "operationId": "LaborController_getCategories",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LaborCategoriesResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "labor"
        ]
      },
      "post": {
        "operationId": "LaborController_createCategory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLaborCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "labor"
        ]
      }
    },
    "/api/labor/categories/{id}": {
      "patch": {
        "operationId": "LaborController_updateCategory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLaborCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "labor"
        ]
      },
      "delete": {
        "operationId": "LaborController_removeCategory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "labor"
        ]
      }
    },
...
  },
  "info": {
    "title": "Auto Core Platform API",
    "description": "Generated OpenAPI spec for contract checks and client types.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreatePurchaseInvoiceDto": {
        "type": "object",
        "properties": {}
      },
...
      "LaborOperationSearchItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "standardAw": {
            "type": "number"
          },
          "hourlyRate": {
            "type": "number"
          },
          "categoryName": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "code",
          "description",
          "standardAw",
          "hourlyRate",
          "categoryName"
        ]
      },
      "LaborOperationSearchMetaDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": [
          "total",
          "limit"
        ]
      },
      "LaborOperationSearchResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LaborOperationSearchItemDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/LaborOperationSearchMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "LaborOperationCategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "LaborOperationFitmentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "make": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "yearFrom": {
            "type": "object",
            "nullable": true
          },
          "yearTo": {
            "type": "object",
            "nullable": true
          },
          "engineCode": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "id",
          "make",
          "model"
        ]
      },
      "LaborOperationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "standardAw": {
            "type": "number"
          },
          "hourlyRate": {
            "type": "number"
          },
          "internalCost": {
            "type": "object",
            "nullable": true
          },
          "categoryId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "category": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/LaborOperationCategoryDto"
              }
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "fitments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LaborOperationFitmentResponseDto"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "code",
          "description",
          "standardAw",
          "hourlyRate",
          "isActive",
          "fitments",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginatedLaborOperationsMetaDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          }
        },
        "required": [
          "total",
          "page",
          "limit",
          "totalPages"
        ]
      },
      "PaginatedLaborOperationsResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LaborOperationResponseDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginatedLaborOperationsMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "LaborOperationFitmentDto": {
        "type": "object",
        "properties": {
          "make": {
            "type": "string",
            "description": "Vehicle make"
          },
          "model": {
            "type": "string",
            "description": "Vehicle model"
          },
          "yearFrom": {
            "type": "number",
            "description": "Year from (inclusive)"
          },
          "yearTo": {
            "type": "number",
            "description": "Year to (inclusive)"
          },
          "engineCode": {
            "type": "string",
            "description": "Engine code"
          }
        },
        "required": [
          "make",
          "model"
        ]
      },
      "CreateLaborOperationDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Unique operation code"
          },
          "description": {
            "type": "string",
            "description": "Operation description"
          },
          "standardAw": {
            "type": "number",
            "description": "Standard allocated work hours (>= 0)",
            "minimum": 0
          },
          "hourlyRate": {
            "type": "number",
            "description": "Hourly rate (> 0)",
            "minimum": 0.01
          },
          "internalCost": {
            "type": "number",
            "description": "Internal cost per operation (overrides category default for cost tracking)"
          },
          "categoryId": {
            "type": "string",
            "description": "Category ID",
            "format": "uuid"
          },
          "isActive": {
            "type": "boolean",
            "description": "Active status",
            "default": true
          },
          "fitments": {
            "description": "Fitments for this operation",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LaborOperationFitmentDto"
            }
          }
        },
        "required": [
          "code",
          "description",
          "standardAw",
          "hourlyRate"
        ]
      },
      "UpdateLaborOperationDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Unique operation code"
          },
          "description": {
            "type": "string",
            "description": "Operation description"
          },
          "standardAw": {
            "type": "number",
            "description": "Standard allocated work hours (>= 0)",
            "minimum": 0
          },
          "hourlyRate": {
            "type": "number",
            "description": "Hourly rate (> 0)",
            "minimum": 0.01
          },
          "internalCost": {
            "type": "number",
            "description": "Internal cost per operation (overrides category default for cost tracking)"
          },
          "categoryId": {
            "type": "string",
            "description": "Category ID",
            "format": "uuid"
          },
          "isActive": {
            "type": "boolean",
            "description": "Active status",
            "default": true
          },
          "fitments": {
            "description": "Fitments for this operation",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LaborOperationFitmentDto"
            }
          }
        }
      },
      "SoftDeleteResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "isActive"
        ]
      },
...
      "LaborCategoryResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "object",
            "nullable": true
          },
          "sort_order": {
            "type": "object",
            "nullable": true
          },
          "parent_id": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "default_hourly_rate": {
            "type": "object",
            "nullable": true
          },
          "is_active": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LaborCategoryChildDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "is_active",
          "createdAt",
          "updatedAt",
          "children"
        ]
      },
      "LaborCategoriesMetaDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "topLevelCount": {
            "type": "number"
          },
          "childCount": {
            "type": "number"
          }
        },
        "required": [
          "total",
          "topLevelCount",
          "childCount"
        ]
      },
      "LaborCategoriesResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LaborCategoryResponseDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/LaborCategoriesMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "CreateLaborCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Unique name for the labor category"
          },
          "description": {
            "type": "string",
            "description": "Description of the category",
            "nullable": true
          },
          "sort_order": {
            "type": "number",
            "description": "Sort order for display purposes"
          },
          "parent_id": {
            "type": "string",
            "description": "Parent category ID (max depth: 1 level)",
            "format": "uuid",
            "nullable": true
          },
          "default_hourly_rate": {
            "type": "number",
            "description": "Default hourly rate for operations in this category",
            "nullable": true
          },
          "is_active": {
            "type": "boolean",
            "description": "Whether the category is active"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateLaborCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Unique name for the labor category"
          },
          "description": {
            "type": "string",
            "description": "Description of the category",
            "nullable": true
          },
          "sort_order": {
            "type": "number",
            "description": "Sort order for display purposes"
          },
          "parent_id": {
            "type": "string",
            "description": "Parent category ID (max depth: 1 level)",
            "format": "uuid",
            "nullable": true
          },
          "default_hourly_rate": {
            "type": "number",
            "description": "Default hourly rate for operations in this category",
            "nullable": true
          },
          "is_active": {
            "type": "boolean",
            "description": "Whether the category is active"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}