# apps/core-api/src/purchase/purchase-invoice.controller.ts
import {
  Controller,
  Get,
  Post,
  Body,
  Param,
  Query,
  Patch,
  Delete,
} from '@nestjs/common';
import { ApiCreatedResponse, ApiOkResponse } from '@nestjs/swagger';
import { PurchaseInvoiceService } from './purchase-invoice.service';
import { CreatePurchaseInvoiceDto } from './dto/create-purchase-invoice.dto';
import { PurchaseInvoiceStatus } from '@prisma/client';

@Controller('purchase-invoices')
export class PurchaseInvoiceController {
  constructor(private readonly service: PurchaseInvoiceService) {}

  @Post()
  @ApiCreatedResponse({
    schema: { type: 'object' },
  })
  create(@Body() createDto: CreatePurchaseInvoiceDto) {
    return this.service.create(createDto);
  }

  @Get()
  @ApiOkResponse({
    schema: { type: 'object' },
  })
  findAll(
    @Query('vendorId') vendorId?: string,
    @Query('status') status?: PurchaseInvoiceStatus,
    @Query('page') page?: string,
    @Query('pageSize') pageSize?: string,
    @Query('sortBy') sortBy: string = 'due_date',
    @Query('order') order: 'asc' | 'desc' = 'asc',
  ) {
    // Validate and normalize pagination parameters
    let pageNum = page ? parseInt(page, 10) : 1;
    let pageSizeNum = pageSize ? parseInt(pageSize, 10) : 25;

    // Ensure valid integers, treating NaN and non-positive values as defaults
    if (!Number.isFinite(pageNum) || pageNum < 1) pageNum = 1;
    if (!Number.isFinite(pageSizeNum) || pageSizeNum < 1) pageSizeNum = 25;

    // Enforce maximum page size
    const MAX_PAGE_SIZE = 100;
    if (pageSizeNum > MAX_PAGE_SIZE) pageSizeNum = MAX_PAGE_SIZE;

    return this.service.findAll(
      vendorId,
      status,
      pageNum,
      pageSizeNum,
      sortBy,
      order,
    );
  }

  @Get(':id')
  @ApiOkResponse({
    schema: { type: 'object' },
  })
  findOne(@Param('id') id: string) {
    return this.service.findOne(id);
  }

  @Patch(':id')
  @ApiOkResponse({
    schema: { type: 'object' },
  })
  update(@Param('id') id: string, @Body() updateDto: CreatePurchaseInvoiceDto) {
    return this.service.update(id, updateDto);
  }

  @Patch(':id/post')
  @ApiOkResponse({
    schema: { type: 'object' },
  })
  post(@Param('id') id: string) {
    return this.service.post(id);
  }

  @Patch(':id/pay')
  @ApiOkResponse({
    schema: { type: 'object' },
  })
  pay(@Param('id') id: string) {
    return this.service.pay(id);
  }

  @Delete(':id')
  remove(@Param('id') id: string) {
    return this.service.remove(id);
  }

  @Delete(':id/lines/:lineId')
  removeLine(@Param('id') id: string, @Param('lineId') lineId: string) {
    return this.service.removeLine(id, lineId);
  }
}